Fix sqlite3_last_insert_rowid return type to long
The long value is converted back to int in HyenaSqliteConnection, to
avoid unnecessary API breakage. Fixes BGO#547218.

+++ src/Libraries/Mono.Data.Sqlite/Mono.Data.Sqlite/SQLite3.cs
@@ -665,7 +665,7 @@ namespace Mono.Data.Sqlite
       }
     }
 
-    internal override int GetLastInsertRowId ()
+    internal override long GetLastInsertRowId ()
     {
 	    return UnsafeNativeMethods.sqlite3_last_insert_rowid (_sql);
     }
